backend/requirements.txt
Django~=4.2.27
djangorestframework~=3.14.0
django-cors-headers~=4.2.0
Pillow~=10.3
python-decouple~=3.8
dj-database-url~=2.0.0
djangorestframework-simplejwt~=5.3.0
psycopg2-binary~=2.9.9
python-dateutil~=2.9.0
whitenoise~=6.6.0
django-filter~=23.5
gunicorn~=21.2.0